monostore/requirements.txt
asgiref==3.9.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
PyMySQL==1.1.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
